Delete 'FileHashStoreReferencesTest' class, move and organize junit tests to 'FileHashStoreProtected', fix incorrect import in 'FileHashStoreInterface' class and use correct 'Disabled' decorator
